Move common files to core folder
