+++ src/bar-chart.js
@@ -2,10 +2,10 @@ import * as d3 from 'd3';
 
 import {stackMixin} from './stack-mixin';
 import {coordinateGridMixin} from './coordinate-grid-mixin';
-import {override, transition} from './core';
-import {constants} from './constants';
-import {logger} from './logger';
-import {pluck, utils} from './utils';
+import {override, transition} from './core/core';
+import {constants} from './core/constants';
+import {logger} from './core/logger';
+import {pluck, utils} from './core/utils';
 
 /**
  * Concrete bar chart/histogram implementation.
