+++ src/base-mixin.js
@@ -1,13 +1,13 @@
 import * as d3 from 'd3';
 
-import { printers, utils } from './utils';
-import { pluck } from './utils';
-import { deregisterChart, instanceOfChart, renderAll, redrawAll, registerChart } from './core';
-import { constants } from './constants';
-import { events } from './events';
-import { logger } from './logger';
-import {InvalidStateException} from './invalid-state-exception';
-import {BadArgumentException} from './bad-argument-exception';
+import { printers, utils } from './core/utils';
+import { pluck } from './core/utils';
+import { deregisterChart, instanceOfChart, renderAll, redrawAll, registerChart } from './core/core';
+import { constants } from './core/constants';
+import { events } from './core/events';
+import { logger } from './core/logger';
+import {InvalidStateException} from './core/invalid-state-exception';
+import {BadArgumentException} from './core/bad-argument-exception';
 
 /**
  * `dc.baseMixin` is an abstract functional object representing a basic `dc` chart object
