+++ src/bubble-mixin.js
@@ -1,8 +1,8 @@
 import * as d3 from 'd3';
 
 import {colorMixin} from './color-mixin';
-import {transition} from './core';
-import {events} from './events';
+import {transition} from './core/core';
+import {events} from './core/events';
 
 /**
  * This Mixin provides reusable functionalities for any chart that needs to visualize data using bubbles.
