+++ src/bubble-overlay.js
@@ -2,9 +2,9 @@ import * as d3 from 'd3';
 
 import {baseMixin} from './base-mixin';
 import {bubbleMixin} from './bubble-mixin';
-import {transition} from './core';
-import {constants} from './constants';
-import {utils} from './utils';
+import {transition} from './core/core';
+import {constants} from './core/constants';
+import {utils} from './core/utils';
 
 /**
  * The bubble overlay chart is quite different from the typical bubble chart. With the bubble overlay
