+++ src/core/bad-argument-exception.js
