+++ src/core/core.js
