+++ src/index.js
@@ -1,13 +1,13 @@
 export * from '../generated/version';
-export * from './constants';
-export * from './core';
-export * from './bad-argument-exception';
-export * from './invalid-state-exception';
-export * from './utils';
-export * from './logger';
-export * from './config';
-export * from './events';
-export * from './filters';
+export * from './core/constants';
+export * from './core/core';
+export * from './core/bad-argument-exception';
+export * from './core/invalid-state-exception';
+export * from './core/utils';
+export * from './core/logger';
+export * from './core/config';
+export * from './core/events';
+export * from './core/filters';
 export * from './d3.box';
 
 export * from './base-mixin';
