+++ src/line-chart.js
@@ -1,10 +1,10 @@
 import * as d3 from 'd3';
 
-import {logger} from './logger';
-import {pluck, utils} from './utils';
+import {logger} from './core/logger';
+import {pluck, utils} from './core/utils';
 import {stackMixin} from './stack-mixin';
 import {coordinateGridMixin} from './coordinate-grid-mixin';
-import {override, transition} from './core';
+import {override, transition} from './core/core';
 
 /**
  * Concrete line/area chart implementation.
