+++ src/pie-chart.js
@@ -2,7 +2,7 @@ import * as d3 from 'd3';
 import {capMixin} from './cap-mixin';
 import {colorMixin} from './color-mixin';
 import {baseMixin} from './base-mixin';
-import {transition} from './core';
+import {transition} from './core/core';
 
 
 /**
